chore: update comments for funcs
